+++ sfAudio/src/SoundFileReaderWav.cpp
@@ -204,7 +204,7 @@ uint64_t SoundFileReaderWav::read(int16_t* samples, uint64_t maxCount)
             {
                 uint32_t sample = 0;
                 if (decode24bit(*m_stream, sample))
-                    *samples++ = sample >> 8;
+                    *samples++ = static_cast<int16_t>(sample >> 8);
                 else
                     return count;
                 break;
